pyproject.toml
[project]
name = "ot-nlp"
version = "0.1.0"
description = "Biblical Hebrew NLP analysis pipeline"
requires-python = ">=3.11,<3.12"
dependencies = [
    "psycopg2-binary==2.9.9",
    "pgvector==0.2.5",
    "pyyaml==6.0.1",
    "text-fabric==12.6.4",
    "numpy==1.26.4",
    "pandas==2.2.2",
    "sphinx==8.1.3",
    "myst-nb==1.3.0",
    "sphinx-book-theme==1.1.3",
    "sphinxcontrib-bibtex==2.6.3",
    "nbconvert==7.16.4",
    "ipykernel==6.29.5",
    "plotly==5.22.0",
    "kaleido==0.2.1",
    "streamlit==1.35.0",
    "usfm-grammar==3.1.2; sys_platform == 'linux'",
    "pronouncing==0.2.0",
    "anthropic==0.28.1",
    "openai==1.35.0",
    "google-generativeai==0.7.2",
    "requests==2.32.3",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2",
    "ruff>=0.4",
    "pyright>=1.1",
    "pre-commit>=3.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
include = ["pipeline", "tests"]
extraPaths = ["pipeline"]
